academia-back/academiab/requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.10
attrs==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
bootstrap4==0.1.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
dj-static==0.0.6
Django==3.1.1
django-autoslug==1.9.8
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-cors-headers==3.5.0
django-jazzmin==2.2.11
django-rest-authtoken==2.1.3
django-widget-tweaks==1.4.8
djangorestframework==3.12.1
entrypoints==0.3
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
future==0.18.2
generics==2.0.0
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==20.0.4
idna==2.10
ipykernel==5.3.4
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
mysqlclient==1.4.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.4
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
pycparser==2.20
pygame==1.9.6
Pygments==2.6.1
PyMySQL==0.10.1
pyparsing==2.4.7
pyrsistent==0.17.0
python-dateutil==2.8.1
pytz==2020.1
pywinpty==0.5.7
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
sqlparse==0.3.1
static3==0.7.0
terminado==0.8.3
testpath==0.4.4
tkintertable==1.3.2
tornado==6.0.4
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
psycopg2==2.8.6
traitlets==4.3.3
ipython==7.16.0